requirements.txt
pynmeagps-1.0.16
geojson
